🐛 fixed event

+++ src/app/components/backoffice/event/event-list/event-list.component.ts
@@ -3,6 +3,7 @@ import { ToastrService } from 'ngx-toastr';
 import { Subscription } from 'rxjs';
 import { Event } from 'src/app/core/models/event';
 import { EventService } from 'src/app/core/services/event.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-event-list',
@@ -13,10 +14,9 @@ export class EventListComponent implements OnInit, OnDestroy {
 
   events: Event[];
   subslist: Subscription
-  subsdele:Subscription
   //contructor khater ytlanca howa o ngOnInit awl haja fl component doub ma thel component yetlanca onstructor mbaaed ngOnit
   constructor(private eventService: EventService,
-    private toasterService:ToastrService) { }
+    private toasterService: ToastrService) { }
 
   //ngOnInt ngOnDestroy method of life cycle compoenent also known as HOOKS
   ngOnInit(): void {
@@ -33,24 +33,39 @@ export class EventListComponent implements OnInit, OnDestroy {
 
   ngOnDestroy(): void {
     this.subslist.unsubscribe();
-    this.subsdele.unsubscribe();
   }
 
   deleteEv(postId: any) {
-    console.log("id",postId);
-
-   this.eventService.deleteEvent(postId).subscribe({
-      complete: () => {
-        this.refresh();
-        this.toasterService.success("success","done!")
-      },
-      error: (err) => {
-        throw err;
+    Swal.fire({
+      title: 'Are you sure want to remove this event ?',
+      text: 'You will not be able to recover this event !',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Yes, delete it!',
+      cancelButtonText: 'No, keep it',
+    }).then((result) => {
+      if (result.value) {
+        this.eventService.deleteEvent(postId).subscribe({
+          complete: () => {
+            this.refresh();
+            Swal.fire(
+              'Deleted!',
+              'Your event has been deleted.',
+              'success'
+            );
+          },
+          error: (err) => {
+            throw err;
+          }
+        });
+      } else if (result.dismiss === Swal.DismissReason.cancel) {
+        Swal.fire('Cancelled', 'Your event is safe :)', 'error');
       }
     });
+
   }
 
-  refresh(){
+  refresh() {
     this.ngOnInit();
   }
 }
